backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
qdrant-client
openai
PyPDF2
python-multipart
python-dotenv
pydantic
langchain-community
langsmith
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy
graphviz